requirements.txt
Django==1.11
Pillow==4.1.0
psycopg2==2.7.1
django-widget-tweaks==1.4.1
django-filter==1.0.2
channels==1.1.3
django-simple-history==1.8.2
